+++ Tests/Unit/Routing/DynamicRouterTest.php
@@ -49,11 +49,10 @@ class DynamicRouterTest extends CmfUnitTestCase
         $this->generator = $this->buildMock('Symfony\\Component\\Routing\\Generator\\UrlGeneratorInterface');
 
         $this->request = Request::create('/foo');
-        $this->container = $this->buildMock('Symfony\\Component\\DependencyInjection\\ContainerInterface');
         $this->context = $this->buildMock('Symfony\\Component\\Routing\\RequestContext');
         $this->eventDispatcher = $this->buildMock('Symfony\Component\EventDispatcher\EventDispatcherInterface');
         $this->router = new DynamicRouter($this->context, $this->matcher, $this->generator, '', $this->eventDispatcher);
-        $this->router->setContainer($this->container);
+        $this->router->setRequest($this->request);
     }
 
     private function assertRequestAttributes($request)
@@ -66,12 +65,6 @@ class DynamicRouterTest extends CmfUnitTestCase
 
     public function testMatch()
     {
-        $this->container->expects($this->once())
-            ->method('get')
-            ->with('request')
-            ->will($this->returnValue($this->request))
-        ;
-
         $this->eventDispatcher->expects($this->once())
             ->method('dispatch')
             ->with(Events::PRE_DYNAMIC_MATCH, $this->equalTo(new RouterMatchEvent()))
@@ -101,11 +94,7 @@ class DynamicRouterTest extends CmfUnitTestCase
      */
     public function testMatchNoRequest()
     {
-        $this->container->expects($this->once())
-            ->method('get')
-            ->with('request')
-            ->will($this->returnValue(null))
-        ;
+        $this->router->setRequest(null);
 
         $this->eventDispatcher->expects($this->once())
             ->method('dispatch')