+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 1.2.0-RC2
 ---------
 
+* **2014-04-14**: DynamicRouter no longer implements the ContainerAwareInterface
 * **2014-04-11**: drop Symfony 2.2 compatibility
 
 1.2.0-RC1